Split response error handling logic
