+++ src/components/response/utils.ts
@@ -2,41 +2,30 @@ import { useForm } from '@mantine/form';
 import { useEffect, useState } from 'react';
 import isEqual from 'lodash.isequal';
 import {
-  CheckboxResponse, CustomResponse, DropdownResponse, JsonValue, MatrixResponse, NumberOption, NumericalResponse, RadioResponse, Response, StringOption,
+  CheckboxResponse, CustomResponse, JsonValue, RadioResponse, Response,
 } from '../../parser/types';
 import { CustomResponseValidate, StoredAnswer } from '../../store/types';
 import { parseStringOptionValue } from '../../utils/stringOptions';
+import {
+  checkCheckboxResponseForValidation,
+  checkMatrixResponse,
+  isEmptyCustomResponseValue,
+  isOtherSelectionIncomplete,
+  REQUIRED_ERROR_MESSAGE,
+  shouldBypassValidationForStandaloneDontKnow,
+  usesStandaloneDontKnowField,
+} from './responseErrors';
 
 type ResponseDefault = JsonValue;
 type ResponseWithDefault = Response & { default?: ResponseDefault };
 
 export const DONT_KNOW_DEFAULT_VALUE = "I don't know";
-export const REQUIRED_ERROR_MESSAGE = 'Please answer this question to continue.';
-export type ResponseIssueType = 'unanswered' | 'invalid';
-export type ResponseIssueSummary = { unansweredCount: number; invalidCount: number };
-
-function isEmptyCustomResponseValue(value: JsonValue | undefined): boolean {
-  if (value === null || value === undefined || value === '') {
-    return true;
-  }
-
-  if (Array.isArray(value)) {
-    return value.length === 0 || value.every((entry) => isEmptyCustomResponseValue(entry));
-  }
-
-  if (typeof value === 'object') {
-    const objectValues = Object.values(value);
-    return objectValues.length === 0 || objectValues.every((entry) => isEmptyCustomResponseValue(entry));
-  }
-
-  return false;
-}
 
 export function normalizeCheckboxDontKnowValue(value: string[]) {
   return value.includes(DONT_KNOW_DEFAULT_VALUE) ? [] : value;
 }
 
-function checkDropdownResponse(dropdownResponse: DropdownResponse, value: string[]) {
+function checkDropdownResponse(dropdownResponse: { minSelections?: number; maxSelections?: number }, value: string[]) {
   // Check max and min selections
   const minNotSelected = dropdownResponse.minSelections && value.length < dropdownResponse.minSelections;
   const maxNotSelected = dropdownResponse.maxSelections && value.length > dropdownResponse.maxSelections;
@@ -50,35 +39,7 @@ function checkDropdownResponse(dropdownResponse: DropdownResponse, value: string
   return null;
 }
 
-function checkCheckboxResponse(response: CheckboxResponse, value: string[]) {
-  const minNotSelected = response.minSelections && value.length < response.minSelections;
-  const maxNotSelected = response.maxSelections && value.length > response.maxSelections;
-
-  if (minNotSelected && maxNotSelected) {
-    return `Please select between ${response.minSelections} and ${response.maxSelections} options`;
-  }
-  if (minNotSelected) {
-    return `Please select at least ${response.minSelections} options`;
-  }
-  if (maxNotSelected) {
-    return `Please select at most ${response.maxSelections} options`;
-  }
-  return null;
-}
-
-export function checkCheckboxResponseForValidation(
-  response: CheckboxResponse,
-  value: string[],
-  dontKnowChecked = false,
-) {
-  if (response.withDontKnow && dontKnowChecked) {
-    return null;
-  }
-
-  return checkCheckboxResponse(response, value);
-}
-
-function checkNumericalResponse(response: NumericalResponse, value: number) {
+function checkNumericalResponse(response: { min?: number; max?: number }, value: number) {
   const numValue = typeof value === 'string' ? parseFloat(value) : value;
 
   const { min, max } = response;
@@ -95,54 +56,6 @@ function checkNumericalResponse(response: NumericalResponse, value: number) {
   return null;
 }
 
-function checkMatrixResponse(response: MatrixResponse, value: Record<string, string>) {
-  const expectedQuestionKeys = response.questionOptions.map((entry) => parseStringOptionValue(entry));
-  const unanswered = expectedQuestionKeys.some((questionKey) => {
-    const rowValue = value[questionKey];
-    return rowValue === undefined || rowValue === '';
-  });
-
-  if (unanswered) {
-    return 'Please answer all questions in the matrix to continue.';
-  }
-
-  return null;
-}
-
-function checkMatrixResponseForMessage(response: MatrixResponse, value: Record<string, string>) {
-  const hasAnsweredAtLeastOne = Object.values(value).some((val) => val !== '');
-  if (!hasAnsweredAtLeastOne) {
-    return null;
-  }
-
-  return checkMatrixResponse(response, value);
-}
-
-function hasOtherText(value: StoredAnswer['answer'][string] | undefined) {
-  return typeof value === 'string' && value.trim().length > 0;
-}
-
-function isOtherSelectionIncomplete(
-  response: Response,
-  value: StoredAnswer['answer'][string],
-  values: StoredAnswer['answer'],
-) {
-  if (!('withOther' in response) || !response.withOther) {
-    return false;
-  }
-
-  const otherInputValue = values[`${response.id}-other`];
-  if (response.type === 'radio') {
-    return value === 'other' && !hasOtherText(otherInputValue);
-  }
-
-  if (response.type === 'checkbox') {
-    return Array.isArray(value) && value.includes('__other') && !hasOtherText(otherInputValue);
-  }
-
-  return false;
-}
-
 const getQueryParameters = () => {
   if (typeof window === 'undefined') {
     return new URLSearchParams('');
@@ -150,15 +63,6 @@ const getQueryParameters = () => {
 
   return new URLSearchParams(window.location.search);
 };
-
-// Matrix questions with "Don't know" option require a separate field to properly handle the "Don't know" state
-export const usesStandaloneDontKnowField = (response: Response) => !!response.withDontKnow
-  && response.type !== 'matrix-radio'
-  && response.type !== 'matrix-checkbox';
-
-export const shouldBypassValidationForStandaloneDontKnow = (response: Response, dontKnowChecked: boolean) => (
-  usesStandaloneDontKnowField(response) && dontKnowChecked
-);
 
 export const getDefaultFieldValue = (response: Response) => {
   const responseDefault = (response as ResponseWithDefault).default;
@@ -456,221 +360,4 @@ export function useAnswerField(
   }, [_id, answerField, currentStep]);
 
   return answerField;
-}
-
-export function generateErrorMessage(
-  response: Response,
-  answer: {
-    value?: number | string | string[] | Record<string, string>;
-    checked?: string[];
-  },
-  options?: (StringOption | NumberOption)[],
-  errorOptions?: {
-    showRequiredErrors?: boolean;
-    values?: StoredAnswer['answer'];
-  },
-) {
-  const { requiredValue, requiredLabel } = response;
-  const showRequiredErrors = !!errorOptions?.showRequiredErrors;
-  const values = errorOptions?.values || {};
-  const dontKnowChecked = !!values[`${response.id}-dontKnow`];
-  const otherValue = values[`${response.id}-other`];
-
-  if (shouldBypassValidationForStandaloneDontKnow(response, dontKnowChecked)) {
-    return null;
-  }
-
-  let error: string | null = '';
-  const checkboxValues = Array.isArray(answer.checked)
-    ? answer.checked
-    : (Array.isArray(answer.value) ? answer.value : undefined);
-  const isEmptyMatrixResponse = typeof answer.value === 'object'
-    && !Array.isArray(answer.value)
-    && answer.value !== null
-    && Object.values(answer.value).every((val) => val === '');
-
-  if (showRequiredErrors && response.required) {
-    if (checkboxValues && checkboxValues.length === 0) {
-      return REQUIRED_ERROR_MESSAGE;
-    }
-
-    if ((response.type === 'matrix-radio' || response.type === 'matrix-checkbox') && isEmptyMatrixResponse) {
-      return REQUIRED_ERROR_MESSAGE;
-    }
-
-    if ((answer.value === null || answer.value === undefined || answer.value === '')
-      && !dontKnowChecked) {
-      return REQUIRED_ERROR_MESSAGE;
-    }
-  }
-
-  if ('withOther' in response && response.withOther) {
-    if (response.type === 'radio' && answer.value === 'other' && !hasOtherText(otherValue)) {
-      return 'Please fill in Other to continue.';
-    }
-
-    if (response.type === 'checkbox' && checkboxValues?.includes('__other') && !hasOtherText(otherValue)) {
-      return 'Please fill in Other to continue.';
-    }
-  }
-
-  if (checkboxValues && Array.isArray(requiredValue)) {
-    error = requiredValue && [...requiredValue].sort().toString() !== [...checkboxValues].sort().toString() ? `Please ${options ? 'select' : 'enter'} ${requiredLabel || requiredValue.toString()} to continue.` : null;
-  } else if (checkboxValues && response.required && response.type === 'checkbox') {
-    error = checkCheckboxResponseForValidation(response, checkboxValues, dontKnowChecked);
-  } else if (answer.value && response.type === 'dropdown') {
-    error = checkDropdownResponse(response, answer.value as string[]);
-  } else if (answer.value && typeof answer.value === 'number' && response.type === 'numerical' && checkNumericalResponse(response, answer.value)) {
-    error = checkNumericalResponse(response, answer.value);
-  } else if (answer.value && typeof answer.value === 'object' && !Array.isArray(answer.value) && (response.type === 'matrix-radio' || response.type === 'matrix-checkbox')) {
-    return checkMatrixResponseForMessage(response, answer.value);
-  } else {
-    error = answer.value && requiredValue && requiredValue.toString() !== answer.value.toString() ? `Please ${options ? 'select' : 'enter'} ${requiredLabel || (options ? options.find((opt) => opt.value === requiredValue)?.label : requiredValue.toString())} to continue.` : null;
-  }
-
-  return error;
-}
-
-// Returns whether a response is unanswered or invalid
-// e.g. for a required question, if no answer is provided, it's unanswered. If an answer is provided but doesn't meet validation criteria, it's invalid.
-export function getResponseIssueType(
-  response: Response,
-  values: StoredAnswer['answer'],
-  customValidate?: CustomResponseValidate,
-  loadError?: string,
-): ResponseIssueType | null {
-  const value = values[response.id];
-  const dontKnowChecked = !!values[`${response.id}-dontKnow`];
-
-  // no need to validate
-  if (response.type === 'textOnly' || response.type === 'divider' || response.type === 'reactive') {
-    return null;
-  }
-
-  if (response.type === 'custom') {
-    if (loadError) {
-      return 'invalid';
-    }
-
-    if (shouldBypassValidationForStandaloneDontKnow(response, dontKnowChecked)) {
-      return null;
-    }
-
-    if (response.required === false && isEmptyCustomResponseValue(value)) {
-      return null;
-    }
-
-    if (isEmptyCustomResponseValue(value)) {
-      return response.required === false ? null : 'unanswered';
-    }
-
-    if (response.requiredValue !== undefined && !isEqual(value, response.requiredValue)) {
-      return 'invalid';
-    }
-
-    if (!customValidate) {
-      return null;
-    }
-
-    return customValidate(value, values, response) ? 'invalid' : null;
-  }
-
-  if (shouldBypassValidationForStandaloneDontKnow(response, dontKnowChecked)) {
-    return null;
-  }
-
-  // If the "Other" option is selected but the accompanying text input is not filled out, consider the response invalid
-  if (isOtherSelectionIncomplete(response, value, values)) {
-    return 'invalid';
-  }
-
-  if (typeof value === 'object' && !Array.isArray(value) && value !== null) {
-    if (response.type === 'matrix-radio' || response.type === 'matrix-checkbox') {
-      const matrixValue = value as Record<string, string>;
-      const hasAnsweredAtLeastOne = Object.values(matrixValue).some((entry) => entry !== '');
-
-      if (!hasAnsweredAtLeastOne) {
-        return response.required ? 'unanswered' : null;
-      }
-
-      // If at least one question in the matrix has been answered, but not all questions have been answered, consider it invalid
-      return checkMatrixResponse(response, matrixValue) ? 'invalid' : null;
-    }
-
-    if (response.type === 'ranking-sublist' || response.type === 'ranking-categorical' || response.type === 'ranking-pairwise') {
-      return Object.keys(value).length === 0 && response.required ? 'unanswered' : null;
-    }
-  }
-
-  if (Array.isArray(value)) {
-    if (value.length === 0) {
-      return response.required ? 'unanswered' : null;
-    }
-
-    if (Array.isArray(response.requiredValue)) {
-      const sortedRequired = [...response.requiredValue].sort();
-      const sortedValue = [...value].sort();
-      const matches = sortedRequired.length === sortedValue.length
-        && sortedRequired.every((entry, idx) => entry === sortedValue[idx]);
-
-      // If there is a requiredValue and the array value doesn't match it, consider it invalid
-      if (!matches) {
-        return 'invalid';
-      }
-    }
-
-    // For checkboxes, if there is a value but it doesn't meet min/max selection criteria, consider it invalid
-    if (response.type === 'checkbox') {
-      return checkCheckboxResponseForValidation(response, value as string[], dontKnowChecked) ? 'invalid' : null;
-    }
-
-    // For dropdowns, if there is a value but it doesn't meet min/max selection criteria, consider it invalid
-    if (response.type === 'dropdown') {
-      return checkDropdownResponse(response, value as string[]) ? 'invalid' : null;
-    }
-
-    return null;
-  }
-
-  if (value === null || value === undefined || value === '') {
-    return response.required ? 'unanswered' : null;
-  }
-
-  // If there is a value but it doesn't match the requiredValue, consider it invalid
-  if (response.requiredValue != null && value.toString() !== response.requiredValue.toString()) {
-    return 'invalid';
-  }
-
-  // For numerical responses, if there is a value but it doesn't meet min/max criteria, consider it invalid
-  if (response.type === 'numerical') {
-    return checkNumericalResponse(response, value as unknown as number) ? 'invalid' : null;
-  }
-
-  return null;
-}
-
-export function summarizeResponseIssues(
-  responses: Response[],
-  values: StoredAnswer['answer'],
-  customResponseValidators: Record<string, CustomResponseValidate | undefined> = {},
-  customResponseLoadErrors: Record<string, string | undefined> = {},
-): ResponseIssueSummary {
-  return responses.reduce<ResponseIssueSummary>((summary, response) => {
-    const issueType = getResponseIssueType(
-      response,
-      values,
-      customResponseValidators[response.id],
-      customResponseLoadErrors[response.id],
-    );
-
-    if (issueType === 'unanswered') {
-      return { ...summary, unansweredCount: summary.unansweredCount + 1 };
-    }
-
-    if (issueType === 'invalid') {
-      return { ...summary, invalidCount: summary.invalidCount + 1 };
-    }
-
-    return summary;
-  }, { unansweredCount: 0, invalidCount: 0 });
 }
